Update page routes
