+++ src/app/donations/page.tsx
@@ -1,58 +1,49 @@
 'use client';
 
 //Hooks
-import { SetStateAction, useState, Dispatch } from 'react';
-
+import { useState, useEffect } from 'react';
 //Components
-import { Button, ImageList } from '@mui/material';
-import DonationCard from '@/components/DonationCard';
+import ProtectedAdminRoute from '@/components/ProtectedAdminRoute';
+import Loader from '@/components/Loader';
+import Donations from '@/components/Donations';
+//Api
+import { getAllDonations } from '@/api/firebase-donations';
+import { addErrorEvent } from '@/api/firebase';
 //Styles
 import '@/styles/globalStyles.css';
 import styles from '@/components/Browse.module.css';
 //Types
 import { Donation } from '@/models/donation';
 
-import DonationDetails from '@/components/DonationDetails';
-import ProtectedAdminRoute from '@/components/ProtectedAdminRoute';
-import AdminCreateDonation from '@/components/AdminCreateDonation';
+const DonationsPage = () => {
+    const [isLoading, setIsLoading] = useState<boolean>(false);
+    const [donations, setDonations] = useState<Donation[] | null>(null);
 
-type DonationsProps = {
-    donations: Donation[];
-    setDonationsUpdated: Dispatch<SetStateAction<boolean>>;
-};
+    async function fetchDonations(): Promise<void> {
+        setIsLoading(true);
+        try {
+            const donationsResult = await getAllDonations();
+            setDonations(donationsResult);
+            setIsLoading(false);
+        } catch (error) {
+            setIsLoading(false);
+            addErrorEvent('Error fetching all donations', error);
+        }
+    }
 
-const Donations = (props: DonationsProps) => {
-    const { donations, setDonationsUpdated } = props;
-    const [idToDisplay, setIdToDisplay] = useState<string | null>(null);
-    const [showForm, setShowForm] = useState<boolean>(false);
-
-    const handleShowForm = () => {
-        //Close details if open
-        setIdToDisplay(null);
-        setShowForm(true);
-    };
+    useEffect(() => {
+        fetchDonations();
+    }, []);
 
     return (
         <ProtectedAdminRoute>
-            {idToDisplay && <DonationDetails id={idToDisplay} setIdToDisplay={setIdToDisplay} setDonationsUpdated={setDonationsUpdated} />}
-            {showForm && <AdminCreateDonation setShowForm={setShowForm} setDonationsUpdated={setDonationsUpdated} />}
-            {!idToDisplay && !showForm && (
-                <>
-                    <div className="page--header">
-                        <h1>Donations</h1>
-                    </div>
-                    <Button variant="contained" type="button" onClick={handleShowForm}>
-                        Add donation
-                    </Button>
-                    <ImageList className={styles['browse__grid']} rowHeight={200}>
-                        {donations.map((donation) => (
-                            <DonationCard key={donation.id} donation={donation} setIdToDisplay={setIdToDisplay} />
-                        ))}
-                    </ImageList>
-                </>
-            )}
+            <>
+                {isLoading && <Loader />}
+                {!isLoading && !donations && <p>No donations found</p>}
+                {!isLoading && donations && <Donations donations={donations} />}
+            </>
         </ProtectedAdminRoute>
     );
 };
 
-export default Donations;
+export default DonationsPage;
