+++ src/app/notifications/page.tsx
@@ -1,90 +1,46 @@
 'use client';
 //Hooks
-import { Dispatch, SetStateAction, useState } from 'react';
+import { Dispatch, SetStateAction, useEffect, useState } from 'react';
 //Components
 import ProtectedAdminRoute from '@/components/ProtectedAdminRoute';
-import UserDetails from '@/components/UserDetails';
-import DonationDetails from '@/components/DonationDetails';
-import NotificationCard from '@/components/NotificationCard';
+import Loader from '@/components/Loader';
+import Notifications from '@/components/Notifications';
 //Api
-
+import { addErrorEvent, getNotifications } from '@/api/firebase';
 //Styles
 import '@/styles/globalStyles.css';
-import styles from '@/components/Browse.module.css';
 //Types
 import { Notification } from '@/types/NotificationTypes';
 
-type NotificationsProps = {
-    notifications: Notification;
-    setNotificationsUpdated: Dispatch<SetStateAction<boolean>>;
-};
+const NotificationsPage = () => {
+    const [isLoading, setIsLoading] = useState<boolean>(false);
+    const [notifications, setNotifications] = useState<Notification | null>(null);
 
-const Notifications = (props: NotificationsProps) => {
-    const { notifications, setNotificationsUpdated } = props;
+    async function fetchNotifications(): Promise<void> {
+        setIsLoading(true);
+        try {
+            const notificationsResult = await getNotifications();
+            setNotifications(notificationsResult);
+            setIsLoading(false);
+        } catch (error) {
+            setIsLoading(false);
+            addErrorEvent('Fetch notifications', error);
+        }
+    }
 
-    const [donationIdToDisplay, setDonationIdToDisplay] = useState<string | null>(null);
-    const [userIdToDisplay, setUserIdToDisplay] = useState<string | null>(null);
-
-    const notificationCount = notifications.donations.length + notifications.users.length;
-    const donationsAwaitingApproval = notifications.donations.filter((donation) => donation.status === 'in processing');
-    const requestedDonations = notifications.donations.filter((donation) => donation.status === 'requested');
-    const usersAwaitingApproval = notifications.users;
+    useEffect(() => {
+        fetchNotifications();
+    }, []);
 
     return (
         <ProtectedAdminRoute>
-            {donationIdToDisplay && <DonationDetails id={donationIdToDisplay} setIdToDisplay={setDonationIdToDisplay} />}
-            {userIdToDisplay && <UserDetails id={userIdToDisplay} setIdToDisplay={setUserIdToDisplay} />}
-            {!donationIdToDisplay && !userIdToDisplay && (
-                <>
-                    <div className="page--header">
-                        <h3>{`You have ${notificationCount} notifications`}</h3>
-                    </div>
-                    {donationsAwaitingApproval.length > 0 && (
-                        <>
-                            <h4>Donations requiring approval</h4>
-                            {donationsAwaitingApproval.map((donation) => (
-                                <NotificationCard
-                                    key={donation.id}
-                                    donation={donation}
-                                    type="pending-donation"
-                                    setIdToDisplay={setDonationIdToDisplay}
-                                    setNotificationsUpdated={setNotificationsUpdated}
-                                />
-                            ))}
-                        </>
-                    )}
-                    {requestedDonations.length > 0 && (
-                        <>
-                            <h4>Requested Equipment</h4>
-                            {requestedDonations.map((donation) => (
-                                <NotificationCard
-                                    key={donation.id}
-                                    type="requested-donation"
-                                    donation={donation}
-                                    setIdToDisplay={setDonationIdToDisplay}
-                                    setNotificationsUpdated={setNotificationsUpdated}
-                                />
-                            ))}
-                        </>
-                    )}
-                    {usersAwaitingApproval.length > 0 && (
-                        <>
-                            <h4>Users awaiting approval</h4>
-                            {usersAwaitingApproval.map((user) => (
-                                <NotificationCard
-                                    key={user.uid}
-                                    type="pending-user"
-                                    authUser={user}
-                                    setIdToDisplay={setUserIdToDisplay}
-                                    setNotificationsUpdated={setNotificationsUpdated}
-                                />
-                            ))}
-                        </>
-                    )}
-                </>
-            )}
+            <>
+                {isLoading && <Loader />}
+                {!isLoading && !notifications && <p>No new notifications</p>}
+                {!isLoading && notifications && <Notifications notifications={notifications} />}
+            </>
         </ProtectedAdminRoute>
     );
 };
 
-export default Notifications;
+export default NotificationsPage;
