+++ src/app/users/page.tsx
@@ -1,71 +1,45 @@
 'use client';
 
 // Hooks
-import React, { Dispatch, SetStateAction, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 // Components
-import { List } from '@mui/material';
-import SearchBar from '@/components/SearchBar';
-import Filter from '@/components/Filter';
-import UserCard from '@/components/UserCard';
-import UserDetails from '@/components/UserDetails';
 import ProtectedAdminRoute from '@/components/ProtectedAdminRoute';
-// Icons
-import { faFilter, faMagnifyingGlass } from '@fortawesome/free-solid-svg-icons';
-import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
-// Styles
-import styles from '@/components/Browse.module.css';
+import Loader from '@/components/Loader';
+import Users from '@/components/Users';
+//Api
+import { addErrorEvent, callListAllUsers } from '@/api/firebase';
+//Styles
 import '@/styles/globalStyles.css';
 // Types
 import { AuthUserRecord } from '@/types/UserTypes';
 
-type UserListProps = {
-    users: AuthUserRecord[];
-    setUsersUpdated: Dispatch<SetStateAction<boolean>>;
-};
+export default function UsersPage() {
+    const [isLoading, setIsLoading] = useState<boolean>(false);
+    const [users, setUsers] = useState<AuthUserRecord[] | null>(null);
 
-export default function Users(props: UserListProps) {
-    const { users, setUsersUpdated } = props;
-    const [isSearchVisible, setIsSearchVisible] = useState<boolean>(false);
-    const [isFilterVisible, setIsFilterVisible] = useState<boolean>(false);
-    const [idToDisplay, setIdToDisplay] = useState<string | null>(null);
-    const [showForm, setShowForm] = useState<boolean>(false);
+    const fetchUsers = async (): Promise<void> => {
+        setIsLoading(true);
+        try {
+            const usersResult = await callListAllUsers();
+            setUsers(usersResult);
+            setIsLoading(false);
+        } catch (error) {
+            setIsLoading(false);
+            addErrorEvent('Error fetching all users', error);
+        }
+    };
 
-    function toggleSearchBar() {
-        setIsSearchVisible((prev: any) => !prev);
-    }
-
-    function toggleFilters() {
-        setIsFilterVisible((prev: any) => !prev);
-    }
+    useEffect(() => {
+        fetchUsers();
+    }, []);
 
     return (
         <ProtectedAdminRoute>
-            {idToDisplay && <UserDetails id={idToDisplay} setIdToDisplay={setIdToDisplay} setUsersUpdated={setUsersUpdated} />}
-
-            {!idToDisplay && !showForm && (
-                <>
-                    <div className={'page--header'}>
-                        <h1>Users</h1>
-                        <div className={styles['header__icons']}>
-                            <div onClick={toggleSearchBar}>
-                                <FontAwesomeIcon icon={faMagnifyingGlass} />
-                            </div>
-                            <div onClick={toggleFilters}>
-                                <FontAwesomeIcon icon={faFilter} />
-                            </div>
-                        </div>
-                    </div>
-                    {isSearchVisible && <SearchBar />}
-                    {isFilterVisible && <Filter />}
-                    <div className="content--container">
-                        <List className={styles['browse__grid']}>
-                            {users.map((userRecord: AuthUserRecord) => {
-                                return <UserCard key={userRecord.uid} authUser={userRecord} setIdToDisplay={setIdToDisplay} />;
-                            })}
-                        </List>
-                    </div>
-                </>
-            )}
+            <>
+                {isLoading && <Loader />}
+                {!isLoading && !users && <p>No users found</p>}
+                {!isLoading && users && <Users users={users} />}
+            </>
         </ProtectedAdminRoute>
     );
 }
